requirements_basic.txt
sentence-transformers
umap-learn
scikit-learn
numpy
pandas
pyarrow
requests
tqdm
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
streamlit
loguru
pyyaml
plotly
